+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (tests) [#11](https://github.com/evmos/proposer/pull/11) Mock API responses for tests.
 - (block) [#15](https://github.com/evmos/proposer/pull/15) Add error handling for blocks data.
 - (crate) [#16](https://github.com/evmos/proposer/pull/16) Adjust package name to `proposer`.
+- (prop) [#22](https://github.com/evmos/proposer/pull/22) Choose GPT model via CLI flag.
 
 ## Legacy Changelog from [Original Repo](https://github.com/MalteHerrmann/upgrade-helper)